Altered the way DRPC is supported by StormCV. Now normal operations can also be used within DRPC topologies without any modification Added a DRPC Example to the examples project to show how it works Further testing is still needed

+++ stormcv/src/main/java/nl/tno/stormcv/util/connector/LocalFileConnector.java
@@ -73,7 +73,7 @@ public class LocalFileConnector implements FileConnector{
 	@Override
 	public void copyFile(File localFile, boolean delete) throws IOException {
 		Files.copy(localFile.toPath(), localLocation.toPath(), StandardCopyOption.REPLACE_EXISTING);
-		localFile.delete();
+		if(delete) localFile.delete();
 	}
 	
 	@Override
